feat: expose track-dist
