+++ index.js
@@ -3,5 +3,6 @@ module.exports = {
 	motp: require('./lib/mot/motp'),
 	motDetails: require('./lib/mot/mot-details'),
 	idDetails: require('./lib/id/id-details'),
+	idTrackDist: require('./lib/id/track-dist'),
 	idf1: require('./lib/id/idf1')
 };
