+++ lib/id/id-details.js
@@ -1,39 +1,8 @@
-const defaultDist = require('../shared/iou-distance');
 const assignement = require('../shared/assignement');
 
-/**
-* List the indexes that are truthy for both
-* @param {Array.<Any>} first
-* @param {Array.<Any>} second
-* @returns {Array.<Number>}
-*/
-const intersection = function (first, second) {
-	const frameIndexes = [first, second]
-		.map(b => b.map((value, index) => ({value, index}))
-			.filter(({value}) => value)
-			.map(({index}) => index));
+const nonNullIndexes = require('./non-null-indexes');
+const trackDist = require('./track-dist');
 
-	return frameIndexes[0].filter(i0 => frameIndexes[1].includes(i0));
-};
-
-const nonNullIndexes = array => {
-	return array
-		.map((value, index) => ({value, index}))
-		.filter(({value}) => value !== null)
-		.map(({index}) => index);
-};
-
-/**
-* @callback DistCb
-* @param {Any} obj1
-* @param {Any} obj2
-* @returns {Number}
-*/
-/**
-* @callback MapCb
-* @param {Any} detection
-* @returns {Any}
-*/
 /**
 * @typedef {Object} IdDetailsResult
 * @property {Number} idtp
@@ -47,26 +16,14 @@ const nonNullIndexes = array => {
 */
 /**
 * List the indexes that are truthy for both
-* @param {Array.<Array.<Any>>} groundTruths
-* @param {Array.<Array.<Any>>} predictions
-* @param {Number} threshold
-* @param {DistCb} [distFn=1 - iou]
-* @param {MapCb} [mapFn = (o => o)]
-* @param {Object} logger console-like (or winston-like) interface for logging
-* @returns {IdDetailsResult}
-*/
-/**
-* List the indexes that are truthy for both
-* @param {Array.<Array.<Any>>} groundTruths
-* @param {Array.<Array.<Any>>} predictions
-* @param {Number} threshold
-* @param {DistCb} [distFn=1 - iou]
-* @param {MapCb} [mapFn = (o => o)]
-* @param {Object} logger console-like (or winston-like) interface for logging
+* @param {TrackDistOptions} opts see trackDist
+* @param {Array.<Array.<Any>>} opts.groundTruths
+* @param {Array.<Array.<Any>>} opts.predictions
+* @param {Object} opts.logger console-like (or winston-like) interface for logging
 * @returns {IdDetailsResult}
 */
 module.exports = function (options) {
-	const {groundTruths, predictions, distFn = defaultDist, threshold = 0.5, mapFn = (o => o), logger} = options;
+	const {groundTruths, predictions, logger} = options;
 	const all = groundTruths.concat(predictions);
 	const nFrames = Math.max(...all.map(a => a.length));
 	// Naming comes from https://arxiv.org/pdf/1609.01775.pdf
@@ -85,32 +42,7 @@ module.exports = function (options) {
 	const vc = groundTruths.map((_, fGtIndex) => ({fGtIndex})).concat(predictions.map((_, predIndex) => ({predIndex})));
 
 	const maxCost = nFrames * 2;
-	/**
-	* @typedef {Object} DetailedDist
-	* @property {Number} fp
-	* @property {Number} fn
-	*/
-	/**
-	* return the number of unmatching {fp, fn} items
-	* @param {Array.<Any>} gt
-	* @param {Array.<Any>} prediction
-	* @returns {DetailedDist}
-	*/
-	const detailedDistFn = function (gt, pred) {
-		const inter = intersection(gt, pred);
-		const spatialMatches = inter.filter(index => distFn(mapFn(gt[index]), mapFn(pred[index])) < threshold);
-		const [fn, fp] = [gt, pred].map(b => (b.filter(a => a).length - spatialMatches.length));
-
-		return {
-			fp,
-			matchIndexes: spatialMatches,
-			fpIndexes: nonNullIndexes(pred).filter(index => !spatialMatches.includes(index)),
-			fnIndexes: nonNullIndexes(gt).filter(index => !spatialMatches.includes(index)),
-			gtL: gt.filter(a => a).length,
-			predL: pred.filter(a => a).length,
-			fn
-		};
-	};
+	const detailedDistFn = (gt, pred) => trackDist.detailed(gt, pred, options);
 
 	/**
 	* If gtIndex and predIndex are both defined, return the number of unmatching (fp + fn) items
