+++ lib/id/non-null-indexes.js
@@ -1,0 +1,6 @@
+module.exports = array => {
+	return array
+		.map((value, index) => ({value, index}))
+		.filter(({value}) => value !== null)
+		.map(({index}) => index);
+};
